+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,3 +1,3 @@
-import { handler } from '../../../../gateways/NextAuthAuthentificationGateway'
+import { handler } from '@/gateways/NextAuthAuthentificationGateway'
 
 export { handler as GET, handler as POST }
