+++ src/components/Gouvernance/FeuilleDeRoute/FeuilleDeRouteRemplie.tsx
@@ -3,12 +3,12 @@
 import Link from 'next/link'
 import { ReactElement, useState } from 'react'
 
-import Table from '../../shared/Table/Table'
 import SectionRemplie from '../SectionRemplie'
 import SubSectionTitle from '../SubSectionTitle'
 import DetailsFeuilleDeRoute from './DetailsFeuilleDeRoute'
 import Drawer from '@/components/shared/Drawer/Drawer'
 import Icon from '@/components/shared/Icon/Icon'
+import Table from '@/components/shared/Table/Table'
 import { FeuilleDeRouteViewModel, GouvernanceViewModel } from '@/presenters/gouvernancePresenter'
 
 export default function FeuilleDeRouteRemplie({
