+++ src/components/shared/Statut/Statut.tsx
@@ -1,6 +1,6 @@
 import { ReactElement } from 'react'
 
-import Badge from '../../shared/Badge/Badge'
+import Badge from '../Badge/Badge'
 
 export default function Statut({ color, libelle }: StatutProps): ReactElement {
   return (
