+++ src/use-cases/commands/InviterUnUtilisateur.test.ts
@@ -1,9 +1,9 @@
 import { InviterUnUtilisateur, InviterUnUtilisateurCommand } from './InviterUnUtilisateur'
 import { EmailGateway } from './shared/EmailGateway'
 import { AddUtilisateurRepository, FindUtilisateurRepository } from './shared/UtilisateurRepository'
-import { TypologieRole } from '../../domain/Role'
-import { Utilisateur, UtilisateurUidState } from '../../domain/Utilisateur'
+import { TypologieRole } from '@/domain/Role'
 import { utilisateurFactory } from '@/domain/testHelper'
+import { Utilisateur, UtilisateurUidState } from '@/domain/Utilisateur'
 
 describe('inviter un utilisateur', () => {
   beforeEach(() => {
